Rolling back AnalyticsParams
